[Fix] Fix padding of IndicatorGraphForPatientScreen and IndicatorMenuForPatientScreen
